Fix scheduler after removal of run_darksky.
